fixed bug/bad merge of files in writing of colors
git-svn-id: file:///pub/import/NoxTools/trunk@74 277ff8d2-6aee-0310-b05b-bd754937ef34

+++ NoxTools/Shared/BinaryIO.cs
@@ -147,14 +147,14 @@ namespace NoxShared
 
 		public void WriteColor(Color color)
 		{
-			Write((short) color.R);
-			Write((short) color.G);
-			Write((short) color.B);
+			Write((byte) color.R);
+			Write((byte) color.G);
+			Write((byte) color.B);
 		}
 
 		public void WriteUserColor(UserColor color)
 		{
-			Write(color.Code - 1);
+			Write((byte) (color.Code - 1));
 		}
 	}
 }
